backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
bcrypt==4.0.1
python-multipart==0.0.6
gunicorn==21.2.0